backend/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ideabase"
version = "0.1.0"
description = "GitHub Trending project analysis and startup ideas API"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi==0.103.1",
    "uvicorn==0.23.2",
    "sqlalchemy==2.0.20",
    "pydantic==2.3.0",
    "alembic==1.12.0",
    "psycopg2-binary==2.9.7",
    "httpx==0.24.1",
    "beautifulsoup4==4.12.2",
    "celery==5.3.4",
    "redis==5.0.0",
    "openai==0.28.1",
    "python-dotenv==1.0.0",
    "pandas==2.1.0",
    "jinja2==3.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.0",
    "pytest-cov==4.1.0",
    "black==23.7.0",
    "isort==5.12.0",
    "mypy==1.5.1",
    "flake8==6.1.0",
]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
